Prevent a "stuck" scroller when attempting to move.
When attempting to move the scroller programatically, if
manualContinuousScrolling is enabled, the scroller will not move if the
scroller is at scrollLeft() == 0 initially (at the first element in the
scroller). I added an initial check to see if swapping should occur
before attempting to move so as to allow the continuous action to occur.

It is not perfect as this will only account for one position worth of
scrolling and attempting to use a step:function during animation seems
to not work because el.data("scrollWrapper").scrollLeft() does not seem
to update properly after swapping so it will scroll for as many steps
in the animation after scrollLeft() == 0.