Added callbacks for jQuery.kinetic (moved/stopped)
Added two callbacks (touchMoved and touchStopped) that are trigged when Smooth Div Scroll is used with the touchScrolling option set to true. These callbacks are just passing on the callbacks that are trigged by jQuery.kinetic (moved and stopped), but are very useful for making things happen when the user interacts with the scroller using touch.