Show both hot spots
This fixes a bug where auto scrolling was being done but one or the other hotspot wasn't showing when it should be. Turns out there is only brief instant during autoscrolling in any direction where both hotspots should not be showing, so this just leaves them on unless the mode is set to always.