Implementing Mishaux's _init function and a new trigger when initialized
When the contents of the widgetized element, or the element itself, are replaced by something other than the widget, we need a way to fully reinitialize the scroller with the new content. This provides an elegant (and I think correct, but I am fairly new to widgets) solution, we just have to recall the widget after the changes are complete.